nlp-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic-settings==2.2.1
python-multipart==0.0.9
openai-whisper==20231117
TTS==0.22.0
langgraph==0.1.5
langchain==0.2.5
langchain-community==0.2.5
ollama==0.2.1
sentence-transformers==2.7.0
faiss-cpu==1.8.0
numpy==1.26.4
